+++ tokogae-data/tokogae-data-event/tokogae-data-event-api/src/main/java/com/tokogae/data/event/service/FoodItemServiceUtil.java
@@ -30,7 +30,7 @@ public class FoodItemServiceUtil {
 	 */
 	public static FoodItem addFoodItem(
 			long subjectId, long occurDay, int occurDaySegment, long occurTime,
-			String name, String locationOfOrigin, String brand, int quantity,
+			String name, String locationOfOrigin, String brand, double quantity,
 			String quantityUnit)
 		throws PortalException {
 
@@ -56,7 +56,7 @@ public class FoodItemServiceUtil {
 
 	public static FoodItem updateFoodItem(
 			long foodItemId, long occurDay, int occurDaySegment, long occurTime,
-			String name, String locationOfOrigin, String brand, int quantity,
+			String name, String locationOfOrigin, String brand, double quantity,
 			String quantityUnit)
 		throws PortalException {
